Move basic driver to private alias
